Rename old implementation to avoid conflict
