+++ pkg/kube_old/global_resource.go
