+++ pkg/kube_old/namespaced_resource.go
