+++ pkg/kube_old/namespaced_resource_dir.go
